went to float8

+++ nodl/classifier/__init__.py
@@ -21,7 +21,7 @@ def relational_classifier(G, labels, epsilon=1e-2, ):
     n_nodes = len(node2id)
 
     # probability matrix
-    P = np.ones((n_nodes, n_labels,), dtype=np.float16) / n_labels
+    P = np.ones((n_nodes, n_labels,), dtype=np.float8) / n_labels
 
     # initialize training set
     for node, label in labels.items():
@@ -46,7 +46,7 @@ def relational_classifier(G, labels, epsilon=1e-2, ):
             end = indptr[node_id + 1]
 
             node_weight = 0
-            probability = np.zeros((n_labels,), dtype=np.float16)
+            probability = np.zeros((n_labels,), dtype=np.float8)
             for neighbor_id, w in zip(indices[start:end], weights[start:end]):
                 probability += w * P[neighbor_id]
                 node_weight += w
